requirements.txt
yfinance>=0.2.36
pandas>=1.5.0
ta>=0.11.0
backtesting>=0.3.3
python-telegram-bot>=21.0
APScheduler>=3.10.4
python-dotenv>=1.0.0
jinja2>=3.1.0
matplotlib>=3.8.0
requests>=2.31.0
